requirement.txt
Django == 1.9
djangorestframework
markdown
django-filter
MySQL-Python
jsonpickle
